Remove organization and region from aws module variables
The AWS provider requires a region and multiple providers are
required to handle multiple regions. Hence, we do not try to outsmart
ourselves here but instead use provider injection into the module
to handle regions as well as organizations (let provider assume a role).

As such, there is no need to pass region or organization into the
cluster and eks modules.