Change cluster ingress to not be nginx specific
To allow users to choose the default ingress, e.g. nginx, haproxy,
or istio's ingress gateway we provision an namespace and a svc
type loadbalancer using TF and integrate this with the IP we reserve
and the DNS zone we provision per cluster.

In the catalog we provide variants for ingress controllers that
set the `kubestack.com/ingress-default: "true"` label, which the svc
selector is set to.

This should give enough flexibility to support multiple use-cases
while giving an easy to use default that works the same across clouds.