Distribute images to speed up automation II
To speed up planning and applying changes under automation we will
build and publish framework images.

This PR changes the release process to include building and
pushing container images and to use these as the FROM in the
Dockerfiles that are part of the quickstart archives.

The Dockerfile's are kept for two reasons:

 1. backwards compatibility, existing automation and instructions
    continue to work
 2. the Dockerfile allows extending the image with additional
    requirements like credential helpers or CAs to validate
    self-signed certificates

The first try at this, using Github Artifacts did not work reliably
enough. The artifact upload would fail silently. Therefor the
artifact approach was reverted. It was replaced with storing
the image and its build-cache in the framework-dev repository.