Handle root -> user differently
Previously the container was started already as non-root. The
purpose of this is to have files created under the mounted volume
match the outside user's user and group id.

Using this approach ssh would complain that a user with the outside
uid and gid did not actually exist inside the container. So
libnss_wrapper was used to fake the user.

This "magic" approach was from a previous Openshift documentation,
but has since been removed there aswell.

Further, to be able to talk to the mounted Docker socket for
the local lab this approach did not work because non-root users
can't change their own groups.

This change creates a user and its groups correctly matching the
host system user and group ids as root. It then executes the auth
helpers and final command using this newly created user.