EKS: Force node_pool to depend on aws-auth configmap fix #114
Using the `eks_node_group` resource the configmap is created by AWS.
This causes an error because it already exists, when the kubernetes
provider tries to create it.

To allow framework users to modify the configmap to control `mapRoles`,
`mapAccounts` and `mapUsers` this commit makes the node_pool module
depend on the configmap.