Extend configuration inheritance implementation
This commit moves the implementation of the configuration
inheritance into a shared module and extends it to support
two commonly requested use-cases.

 1. Support for custom names instead of apps and ops, fix #126
 2. Support for more than two infrastructure environments

The module expects two input variables, `configuration` and
`base_key`.

`configuration` is a map of maps, and `base_key` sets the key
in the configuration map that all other keys will inherit from.

So, with the frameworks default apps and ops inheritance and
configuration, it has the exact same behavior as before. Ops
inherits from apps and overwrites values.

But given a Terraform configuration with different and more keys:

```
locals {
  cfg = {
    "a" = {
      "a" = "a"
      "b" = null
    }
    "b" = {
      "b" = "b"
    }
    "c" = {
      "c" = "c"
    }
  }
}

module "configuration" {
  source = "../common/configuration"

  configuration = local.cfg
  base_key      = "a"
}

output "merged_cfg" {
  value = module.configuration.merged
}
```

The merged_cfg output will be:

```
merged_cfg = {
  "a" = {
    "a" = "a"
  }
  "b" = {
    "a" = "a"
    "b" = "b"
  }
  "c" = {
    "a" = "a"
    "c" = "c"
  }
}
```

Until futher notice the framework documents and tests
only the apps and ops case. So, here be dragons, you've been warned.

But users can now adjust environment names and number to their liking.
Provided they also align the the tagging strategy and triggers
in the pipeline accordingly.