Vendor Terraform providers in container image
This commit extends the Dockerfile and pipeline to vendor Terraform
providers. This change decreases the time a pipeline run takes and
reduces the risk of pipeline runs failing due to unavailability of
external services.

This both improves developer experience and automation reliability.

Terraform will transparently download provider versions that are
not cached.

Unfortunately similar vendoring isn't available for the modules.
This means the modules are still downloaded during every pipeline
run. Goal is to address this in the future.