Fix AKS vnet names #170
The vnet name did not use the metadata_name var, and the code
used to generate the name did only include the workspace.

This meant it supported multiple environments in the same Azure
resource group, but not multiple clusters per environment in
the same resource group.

With this commit, the vnet uses the metadata_name variable and
the subnet will include the name of the node pool.

Since changing the vnet and subnet names would require a
destroy and re-create plan, existing users can keep the old
names by setting `legacy_vnet_name = true`.