Fix configuration module inheritance #176
The [change to support maps of objects][1], not just maps of strings,
introduced a regression. Keys not set in the base environment,
but set in an environment that inherits from base were lost.

Previously ops inherited from apps and could overwrite values
set in apps as well as add new key value pairs to the ops env.

This commit fixes the regression by looping through a set of all
keys from both the environments to be merged.

[1]: https://github.com/kbst/terraform-kubestack/commit/22b0eb8c9439d8f4f2480c04f59acfeba54e6be9