Stop vendoring providers
While vendoring providers is great to make the automation runs
more robust against third party services being unavailable, and
to keep the runtime of each run low, it does not align well with
the direction Terraform upstream is taking with the new lock file.

With vendored providers inside the Kubestack container, Terraform
by default does not check the registry for new versions. Users
have to explicitly enable that using a .terraformrc file. Once they
do, Terraform starts downloading new versions, if available and
specified in the lock file. Leaving them downloading the old
vendored versions as part of the container and still downloading
newer versions during init.

If users do not explicitly enable the registry via the .terraformrc
file, updating the lock file using terraform init --upgrade
requires users to be aware of the vendored version in the container.

At this point I decided to remove the current provider vendoring
approach, and just follow upstream with the lock file. If a better
solution for vendoring is available, optimally for providers and
modules togehter, I will revisit this feature.