Speed up dist with smaller dist-helper stage
Caching to a registry image requires running the build in a
container using buildkit. To be able to run the image and
copy the artifacts out of the container using docker cp
the image needs to be exported from the build container to
the host.

This additional build stage means, the image that is exported
to the host is much smaller by not including the Python and
Terraform caches.