EKS: Remove depends_on from eks_zero_dns_zone
The depends_on propagates to the data sources in the module
and causes the Route53 zone data source to only be read on
apply. This means the zone_id referenced in the Route53 host
resources will be in known-after-apply, which causes the
hosts to be recreated unnecessarily. Since this would cause
small time windows where uncached DNS queries would fail,
not having the explicit depends_on is the better compromise.
Even if it adds another reason why --target apply is required
during bootstrap or recreation.