Deprecate managing the aws-auth config map in the cluster module
Managing the aws-auth config map requires a configured K8s
provider, which frequently requires --target applies,
when attributes of the cluster resource are in known-after-apply
state and therefor the provider can't be configured, leading to
permission denied or connection refused errors.

Further, the config map is created automatically, when the first
managed node pool is created. Which previously required a rather
complex depends_on setup to make sure Terraform creates the
config map before the EKS control plane does.

Finally, the combination of the in-module provider (a Terraform
anti-pattern) and the complex depends_on meant that the --target
apply was further complicated by having to manually target the
cluster resource.

This PR deprecated the ability to manage the aws-auth config map
through the cluster module in favor of using K8s RBAC directory.

The tests/eks_zero_rbac.tf shows the replacement for the
previously used cluster_aws_auth_map_users.