Revert configurable VPC DNS settings to previous default
Before making this configurable the defaults were DNS hostnames
disabled and DNS support enabled. With the defaults set in the
previous PR the nodes are not able to join the cluster. DNS
support needs to be enabled.

I also removed the defaults from the variables. I believe it's
better to set the defaults only in one place.