Allow setting AKS ingress IP zones
Upstream azurerm refactored how zones are handled
across resources. Kubestack let Azure handle the
zones for the ingress IP, but previous provider
versions stored them in state.

The upgraded provider wants to destory and recreate
the IP because the zones in state don't match what's
specified in code (null).

Users that do not want the IP to be recreated, have
to set the zones explicityly to match what's in
state. This commit adds the configuration attribute
to do so.