Do not set any service CIDR unless user sets it
The approach I'm taking here leaves the attribute null.

If not set by the user, whatever EKS uses as the default
will be used. If the user overwrites it, their value will
be used.

I prefer this approach, because I do not have to concern
myself at all with whatever the EKS default is.